Move libvirt provider to a specific directory
Mainly just code movement:

1. Libvirt provider is now a module at lago/providers/libvirt.

2. The XML templates now reside in: lago/providers/libvirt/templates.

3. Changed the XML reading mechanism to use pkg_resources, as recommended.

4. Moved the libvirt network class to lago/providers/libvirt/network.py
from virt.py
