+++ lago/providers/libvirt/templates/dom_template.xml
