+++ lago/providers/libvirt/templates/net_br_template.xml
