+++ lago/providers/libvirt/templates/net_nat_template.xml
