+++ lago/providers/libvirt/templates/snapshot_template.xml
