+++ setup.cfg
@@ -66,7 +66,7 @@ lago.plugins.vm =
     ovirt-he-host=ovirtlago.virt:HEHostVM
 
 lago.plugins.vm_provider =
-    local-libvirt=lago.vm:LocalLibvirtVMProvider
+    local-libvirt=lago.providers.libvirt.vm:LocalLibvirtVMProvider
     ssh=lago.vm:SSHVMProvider
 
 lago.plugins.vm_service =
